#135 refactor: split livewire component and view files to separate edit and create, and create basic component and view file
