+++ config/ehealth.php
@@ -23,7 +23,7 @@ return [
         'cooldown' => 300,
         'retries' => 10
     ],
-    
+
     'capitation_contract_max_period_days' => 366,
     'legal_entity_type' => [
         'PRIMARY_CARE' => [
@@ -112,8 +112,7 @@ return [
         ],
         'MED_ADMIN' => [
             'position' => [
-                'P5', 'P6', 'P7', 'P8', 'P9', 'P10', 'P11', 'P12', 'P13', 'P33', 'P34', 'P35', 'P36', 'P37', 'P38',
-                'P39',
+                'P5', 'P6', 'P7', 'P8', 'P9', 'P10', 'P11', 'P12', 'P13', 'P33', 'P34', 'P35', 'P36', 'P37', 'P38', 'P39',
                 'P40', 'P41', 'P42', 'P43', 'P44', 'P45', 'P46', 'P47', 'P48', 'P49', 'P50', 'P51', 'P52', 'P53', 'P54',
                 'P55', 'P56', 'P57', 'P58', 'P59', 'P60', 'P61', 'P62', 'P63', 'P64', 'P65', 'P66', 'P67', 'P68', 'P69',
                 'P70', 'P71', 'P72', 'P73', 'P74', 'P75', 'P76', 'P77', 'P78', 'P79', 'P80', 'P81', 'P82', 'P83', 'P84',
